+++ src/stats.rs
@@ -1,8 +1,8 @@
 use std::sync::atomic::{AtomicU64, Ordering::Relaxed};
 
-/// Contains statistics related to the node.
+/// Contains statistics related to a node.
 #[derive(Default)]
-pub struct NodeStats {
+pub struct Stats {
     /// The number of all messages sent.
     msgs_sent: AtomicU64,
     /// The number of all messages received.
@@ -11,9 +11,11 @@ pub struct NodeStats {
     bytes_sent: AtomicU64,
     /// The number of all bytes received.
     bytes_received: AtomicU64,
+    /// The number of failures.
+    failures: AtomicU64,
 }
 
-impl NodeStats {
+impl Stats {
     /// Registers a sent message of the provided `size` in bytes.
     pub fn register_sent_message(&self, size: usize) {
         self.msgs_sent.fetch_add(1, Relaxed);
@@ -24,6 +26,11 @@ impl NodeStats {
     pub fn register_received_message(&self, size: usize) {
         self.msgs_received.fetch_add(1, Relaxed);
         self.bytes_received.fetch_add(size as u64, Relaxed);
+    }
+
+    /// Registers a failure.
+    pub fn register_failure(&self) {
+        self.failures.fetch_add(1, Relaxed);
     }
 
     /// Returns the number of sent messages and their collective size in bytes.
@@ -40,5 +47,10 @@ impl NodeStats {
         let bytes = self.bytes_received.load(Relaxed);
 
         (msgs, bytes)
+    }
+
+    /// Returns the number of failures.
+    pub fn failures(&self) -> u64 {
+        self.failures.load(Relaxed)
     }
 }